+++ rust/crd/src/lib.rs
@@ -673,7 +673,6 @@ mod tests {
         spec:
           image:
             productVersion: 2.6.1
-            stackableVersion: 0.0.0-dev
           clusterConfig:
             executor: KubernetesExecutor
             loadExamples: true
@@ -723,7 +722,6 @@ mod tests {
         spec:
           image:
             productVersion: 2.6.1
-            stackableVersion: 0.0.0-dev
           clusterConfig:
             executor: CeleryExecutor
             loadExamples: false
@@ -770,7 +768,6 @@ mod tests {
         spec:
           image:
             productVersion: 2.6.1
-            stackableVersion: 0.0.0-dev
           clusterConfig:
             executor: CeleryExecutor
             loadExamples: false