+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Implement KubernetesExecutor stackableVersion to operator version ([#311]).
 - Default stackableVersion to operator version ([#312]).
 
 ### Changed
@@ -14,6 +15,7 @@
 
 [#303]: https://github.com/stackabletech/airflow-operator/pull/303
 [#308]: https://github.com/stackabletech/airflow-operator/pull/308
+[#311]: https://github.com/stackabletech/airflow-operator/pull/311
 [#312]: https://github.com/stackabletech/airflow-operator/pull/312
 
 ## [23.7.0] - 2023-07-14