+++ CHANGELOG.md
@@ -4,6 +4,22 @@
 
 ### Added
 
+- Default stackableVersion to operator version ([#312]).
+
+### Changed
+
+- [BREAKING] Consolidated `spec.clusterConfig.authenticationConfig` to `spec.clusterConfig.authentication` which now takes a vector of AuthenticationClass references  ([#303]).
+- `vector` `0.26.0` -> `0.31.0` ([#308]).
+- `operator-rs` `0.44.0` -> `0.45.1` ([#308]).
+
+[#303]: https://github.com/stackabletech/airflow-operator/pull/303
+[#308]: https://github.com/stackabletech/airflow-operator/pull/308
+[#312]: https://github.com/stackabletech/airflow-operator/pull/312
+
+## [23.7.0] - 2023-07-14
+
+### Added
+
 - Generate OLM bundle for Release 23.4.0 ([#270]).
 - Fix LDAP tests for Openshift ([#270]).
 - Missing CRD defaults for `status.conditions` field ([#277]).
@@ -12,10 +28,14 @@
 - Operator errors out when credentialsSecret is missing ([#293]).
 - Support podOverrides ([#295]).
 
+### Fixed
+
+- Increase the size limit of the log volume ([#299]).
+
 ### Changed
 
 - [BREAKING] Consolidated remaining top-level config options to `clusterConfig` ([#271]).
-- `operator-rs` `0.40.2` -> `0.41.0` ([#272]).
+- `operator-rs` `0.40.2` -> `0.44.0` ([#272], [#299]).
 - Use 0.0.0-dev product images for testing ([#274])
 - Use testing-tools 0.2.0 ([#274])
 - Added kuttl test suites ([#291])
@@ -30,6 +50,7 @@
 [#291]: https://github.com/stackabletech/airflow-operator/pull/291
 [#293]: https://github.com/stackabletech/airflow-operator/pull/293
 [#295]: https://github.com/stackabletech/airflow-operator/pull/295
+[#299]: https://github.com/stackabletech/airflow-operator/pull/299
 
 ## [23.4.0] - 2023-04-17
 
@@ -83,7 +104,9 @@
 - `operator-rs` `0.27.1` -> `0.30.1` ([#208])
 - `operator-rs` `0.30.1` -> `0.31.0` ([#216]).
 - Updated stackable image versions ([#193]).
-- [BREAKING] Use Product image selection instead of version. `spec.version` has been replaced by `spec.image` ([#206]).
+- [BREAKING] Use Product image selection instead of version ([#206]).
+  - `spec.version` has been replaced by `spec.image`.
+  - `spec.statsdExporterVersion` has been removed, the statsd-exporter is now part of the images itself
 - Fixed the RoleGroup `selector`. It was not used before. ([#208])
 - Refactored LDAP related code to use new `LdapAuthenticationProvider` functionality ([#216])